Reorganize Folder structure
Made folder structure nice and clear and scalable
