+++ pages/index.js
@@ -1,0 +1,3 @@
+import Index from '../src/views/pages/index';
+
+export default Index;
