+++ postcss.config.js
@@ -1,7 +1,5 @@
 const autoprefixer = require('autoprefixer');
 
 module.exports = {
-  plugins: [
-    autoprefixer
-  ]
-}
+  plugins: [autoprefixer]
+};
