+++ src/libraries/Velocity.js
