+++ src/views/components/Hero.js
