+++ src/views/components/__body.js
