+++ src/views/pages/index/index.js
@@ -1,0 +1,14 @@
+import WhoWeAre from './WhoWeAre';
+import UpcomingProjects from './UpcomingProjects';
+import News from './News';
+import IndexContainer from './IndexContainer';
+
+const Index = () => (
+  <IndexContainer>
+    <WhoWeAre />
+    <UpcomingProjects />
+    <News />
+  </IndexContainer>
+);
+
+export default Index;
