+++ src/views/templates/HomeTemplate.js
